refactor: consolidate canvas YAML serialization in pkg/canvas/yaml
Canvas YAML logic was split across pkg/canvas/yaml (parsing) and
pkg/grpc/actions/canvases (serialization). Move the canvas.yaml
serialization (CanvasResourceYAML) and its node/edge default and
position-key helpers into pkg/canvas/yaml so a single package owns
canvas.yaml read and write. CanvasResourceYAML takes the already
serialized proto version, keeping the YAML package free of model and
database dependencies. workflow_yaml.go now only holds thin gRPC
adapters that map to canvasyaml/models and wrap errors with status codes.

Addresses review feedback that YAML logic was spread across three
packages.
